+++ apps/backend/backend_plugin.c
@@ -78,19 +78,26 @@ clixon_plugin_reset_one(clixon_plugin_t *cp,
 			clicon_handle  h,
 			char         *db)
 {
-    int          retval = -1;
-    plgreset_t  *fn;       /* callback */
+    int               retval = -1;
+    plgreset_t       *fn;       /* callback */
+    plugin_context_t *pc = NULL;
 
     if ((fn = clixon_plugin_api_get(cp)->ca_reset) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, db) < 0) {
 	    if (clicon_errno < 0) 
 		clicon_log(LOG_WARNING, "%s: Internal error: Reset callback in plugin: %s returned -1 but did not make a clicon_err call",
 			   __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
     
@@ -108,6 +115,7 @@ clixon_plugin_reset_all(clicon_handle h,
     int             retval = -1;
     clixon_plugin_t  *cp = NULL;
     
+    clicon_debug(1, "%s", __FUNCTION__);
     /* Loop through all plugins, call callbacks in each */
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if (clixon_plugin_reset_one(cp, h, db) < 0)
@@ -130,8 +138,12 @@ clixon_plugin_pre_daemon_one(clixon_plugin_t *cp,
 {
     int          retval = -1;
     plgdaemon_t *fn;          /* Daemonize plugin callback function */
+    plugin_context_t   *pc = NULL;
 
     if ((fn = clixon_plugin_api_get(cp)->ca_pre_daemon) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
+
         if (fn(h) < 0) {
             if (clicon_errno < 0)
                 clicon_log(LOG_WARNING, "%s: Internal error: Pre-daemon callback in plugin:\
@@ -139,9 +151,13 @@ clixon_plugin_pre_daemon_one(clixon_plugin_t *cp,
                            __FUNCTION__, clixon_plugin_name_get(cp));
             goto done;
         }
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -160,6 +176,7 @@ clixon_plugin_pre_daemon_all(clicon_handle h)
     int            retval = -1;
     clixon_plugin_t *cp = NULL;
 
+    clicon_debug(1, "%s", __FUNCTION__);
     /* Loop through all plugins, call callbacks in each */
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
         if (clixon_plugin_pre_daemon_one(cp, h) < 0)
@@ -182,17 +199,24 @@ clixon_plugin_daemon_one(clixon_plugin_t *cp,
 {
     int          retval = -1;
     plgdaemon_t *fn;          /* Daemonize plugin callback function */    
+    plugin_context_t   *pc = NULL;
 
     if ((fn = clixon_plugin_api_get(cp)->ca_daemon) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h) < 0) {
 	    if (clicon_errno < 0) 
 		clicon_log(LOG_WARNING, "%s: Internal error: Daemon callback in plugin: %s returned -1 but did not make a clicon_err call",
 			   __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
     
@@ -213,6 +237,7 @@ clixon_plugin_daemon_all(clicon_handle h)
     int            retval = -1;
     clixon_plugin_t *cp = NULL;
     
+    clicon_debug(1, "%s", __FUNCTION__);
     /* Loop through all plugins, call callbacks in each */
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if (clixon_plugin_daemon_one(cp, h) < 0)
@@ -258,10 +283,12 @@ clixon_plugin_statedata_one(clixon_plugin_t *cp,
     int              retval = -1;
     plgstatedata_t  *fn;          /* Plugin statedata fn */
     cxobj           *x = NULL;
+    plugin_context_t *pc = NULL;
     
-    clicon_debug(1, "%s %s", __FUNCTION__, clixon_plugin_name_get(cp));
     if ((fn = clixon_plugin_api_get(cp)->ca_statedata) != NULL){
 	if ((x = xml_new(DATASTORE_TOP_SYMBOL, NULL, CX_ELMNT)) == NULL)
+	    goto done;
+	if ((pc = plugin_context_get()) == NULL)
 	    goto done;
 	if (fn(h, nsc, xpath, x) < 0){
 	    if (clicon_errno < 0) 
@@ -269,11 +296,15 @@ clixon_plugin_statedata_one(clixon_plugin_t *cp,
 			   __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto fail;  /* Dont quit here on user callbacks */
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     if (xp && x)
 	*xp = x;
     retval = 1;
  done:
+    if (pc)
+	free(pc);
     return retval;
  fail:
     retval = 0;
@@ -406,14 +437,20 @@ clixon_plugin_lockdb_one(clixon_plugin_t *cp,
 {
     int          retval = -1;
     plglockdb_t *fn;          /* Plugin statedata fn */
+    plugin_context_t *pc = NULL;
     
-    clicon_debug(1, "%s %s", __FUNCTION__, clixon_plugin_name_get(cp));
     if ((fn = clixon_plugin_api_get(cp)->ca_lockdb) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, db, lock, id) < 0)
 	    goto done;  
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -576,17 +613,24 @@ plugin_transaction_begin_one(clixon_plugin_t    *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_begin) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -605,6 +649,7 @@ plugin_transaction_begin_all(clicon_handle       h,
     int            retval = -1;
     clixon_plugin_t *cp = NULL;
 
+    clicon_debug(1, "%s", __FUNCTION__);
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if (plugin_transaction_begin_one(cp, h, td) < 0)
 	    goto done;
@@ -628,17 +673,24 @@ plugin_transaction_validate_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
-    
+    plugin_context_t *pc = NULL;
+
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_validate) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -679,17 +731,24 @@ plugin_transaction_complete_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_complete) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -762,17 +821,24 @@ plugin_transaction_commit_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_commit) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -821,17 +887,24 @@ plugin_transaction_commit_done_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_commit_done) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -872,17 +945,24 @@ plugin_transaction_end_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_end) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -899,6 +979,7 @@ plugin_transaction_end_all(clicon_handle h,
     int            retval = -1;
     clixon_plugin_t *cp = NULL;
 
+    clicon_debug(1, "%s", __FUNCTION__);
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if (plugin_transaction_end_one(cp, h, td) < 0)
 	    goto done;
@@ -915,17 +996,24 @@ plugin_transaction_abort_one(clixon_plugin_t      *cp,
 {
     int         retval = -1;
     trans_cb_t *fn;
+    plugin_context_t *pc = NULL;
     
     if ((fn = clixon_plugin_api_get(cp)->ca_trans_abort) != NULL){
+	if ((pc = plugin_context_get()) == NULL)
+	    goto done;
 	if (fn(h, (transaction_data)td) < 0){
 	    if (!clicon_errno) /* sanity: log if clicon_err() is not called ! */
 		clicon_log(LOG_NOTICE, "%s: Plugin '%s' callback does not make clicon_err call on error", 
 		       __FUNCTION__, clixon_plugin_name_get(cp));
 	    goto done;
 	}
+	if (plugin_context_check(pc, clixon_plugin_name_get(cp), __FUNCTION__) < 0)
+	    goto done;
     }
     retval = 0;
  done:
+    if (pc)
+	free(pc);
     return retval;
 }
 
@@ -942,6 +1030,7 @@ plugin_transaction_abort_all(clicon_handle       h,
     int            retval = -1;
     clixon_plugin_t *cp = NULL;
 
+    clicon_debug(1, "%s", __FUNCTION__);
     while ((cp = clixon_plugin_each(h, cp)) != NULL) {
 	if (plugin_transaction_abort_one(cp, h, td) < 0)
 	    ; /* dont abort on error */