+++ apps/backend/backend_plugin_restconf.c
@@ -44,6 +44,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <errno.h>
+#include <signal.h>
 #include <assert.h>
 #include <sys/stat.h>