+++ apps/backend/clixon_backend_transaction.c
@@ -51,6 +51,7 @@
 #include <ctype.h>
 #include <sys/types.h>
 #include <regex.h>
+#include <signal.h>
 #include <netinet/in.h>
 #include <limits.h>